[mexOS][allFiles] bugfixes, new features, more debug output....

+++ kernel/include/types.h
@@ -16,7 +16,8 @@ typedef uint32_t           size_t;
 typedef int32_t            ssize_t;
 typedef int32_t            pid_t;
 typedef uint32_t           tid_t;
-typedef uint64_t           uintptr_t;
+//typedef uint64_t           uintptr_t;
+typedef uint32_t           uintptr_t;
 
 /**
  * @brief NULL pointer definition
